Disable config drive and add ssh key to baremetal instances
Having a config drive attached to the baremetal instances breaks
the overcloud deploy, so explicitly disable it in the Heat template
and leave a note about Nova configuration too.

Also, for the sake of debugging the above problem, pass the user's
ssh key to the baremetal instances as well so if they do boot off
config drive metadata it will be possible to connect to them for
debugging.