Don't connect baremetal vms to private network
It turns out that connecting the baremetal vms to a network with
DHCP enabled causes a race condition where they may get cloud-init
metadata from either the undercloud or the nethercloud.

While it's unrealistic to have a deployed system connected only to
the deployment network, for the moment it gets things working. In
the long run, a separate network with DHCP disabled should probably
be created to simulate the "public" network on the overcloud.