Validate vlans used
Prevent multiple networks using the same vlan, but only if those
networks are actually in use.

Also adds a mapping of the lower-case network names to the
camel-case version.  There are various places where we need to be
able to convert from one to the other, including this vlan check,
so it made sense to do it once and reuse.