Add another layer of nesting
This was probably the right way to do bonded interfaces in the
first place, and it's also going to be helpful for adding routes
to various items, including vlans that may already be at the
deepest level of the previous UI architecture.

Note that this is an incompatible data format change, so I've
bumped the major version.  If this were an in-production tool it
would probably have been possible to migrate old data, but at this
point it's not worth it.