Don't assume vlans have a device field
If a vlan is not in a bond, it won't have a device field and the
template generation would fail.  This makes the vlan processing
work correctly in that case.

This breakage was missed because the vlan on a non-bonded bridge
case was missing in the unit tests, so a test config for that case
is also added.