Only default one nic in a bond to primary
Previously all interfaces defaulted to being primary.  This is not
allowed in a bond, so this change flips primary to False on all
interfaces after the first.

Note that there is the potential for this to result in no interfaces
being set to primary by default if a route gets added to the bond
first before any nics.  I don't expect this to be the typical use
case though, and all it requires is the user to manually set the
correct nic to primary if it does happen.