Order dicts in a human-readable way
It's extremely difficult to read the generated yaml when it is
sorted by key.  Members of a single dict can be split by dozens of
lines if the dict has a lot of child entries.  This change forces
dicts to be output in a more readable order.

Note that this is being done at generation time instead of while
building the data for two reasons:
1) This way it is trivial for existing templates to take advantage
   of the improvement.  They just need to be re-generated with a
   new version of the tool.
2) A fair amount of manipulation of the raw data happens during the
   generation process, so the ordering would likely get messed up
   if we don't set the ordering after that manipulation.