Also write network-isolation file with absolute paths
In many cases the packaged templates in
/usr/share/openstack-tripleo-heat-templates are going to be used
with the generated templates.  In this case, it is simpler to have
a network-isolation.yaml with the paths hard-coded to use that
location instead of relative paths that require the file to be
copied into the tht tree.

This absolute file is added as network-isolation-absolute.yaml in
addition to the previously generated file so either can be used as
desired.