Hide rather than disable
Previously some UI elements were just set disabled when they were
not applicable to the selected item.  It's cleaner to completely
hide them, so let's do that.

This will also allow us to add more inputs for different types of
interfaces and such without cluttering up the UI to an unmanageable
degree.